implement and test set_dims
